feat: add SimpleX model and enhance dataset with rating info
- Implement SimpleX model and `UserBehaviorAggregator` in candidate generation project
- Add `AveragePoolingIgnoringPadding` and `IdEmbeddingBag` modules
- Refactor `gSASRecLoss` to `gBCE` and move to `models/loss` directory
- Update `UserTower` and `ItemTower` to support identity hidden layers when dimensions are empty
- Enrich Amazon Reviews dataset with `average_rating` and `rating_number` fields in `ml_sandbox_libs`
- Update `AmazonReviewsSeqRecDataset` to include rating metadata in items and history sequences
- Adjust hyperparameters for DIN model (learning rate, embedding dims) in ranking project
- Remove debug batch limits in ranking training script
- Upgrade `uv` version in `mise.toml`

+++ projects/recsys-candidate-generation/src/my_types/__init__.py
@@ -18,6 +18,6 @@ class LRSchedulerParams:
     frequency: int
     t_initial: int
     warmup_t: int
-    warmup_lr_init: float
+    warmup_lr_init: int
     lr_min: float
     cycle_limit: int
